% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Add.R
\name{Add}
\alias{Add}
\title{Add constructor yet to be implemented}
\usage{
Add(...)
}
\arguments{
\item{...}{fields yet to be implemented}
}
\value{
the object
}
\description{
Add constructor yet to be implemented
}
\examples{
#no example yet
}
