% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{addInstruction<-}
\alias{addInstruction<-}
\title{Add one or more instructions to a Dockerfile}
\usage{
addInstruction(dockerfileObject) <- value
}
\arguments{
\item{dockerfileObject}{An object of class 'Dockerfile'}

\item{value}{An object that inherits from class 'instruction' or a list of instructions}
}
\value{
Returns the modified Dockerfile object (replacement method)
}
\description{
Add one or more instructions to a Dockerfile
}
\examples{
the_dockerfile <- dockerfile(clean_session())
addInstruction(the_dockerfile) <- Label(myKey = "myContent")
}
