% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Arg.R
\name{Arg}
\alias{Arg}
\title{Arg constructor yet to be implemented}
\usage{
Arg(...)
}
\arguments{
\item{...}{fields yet to be implemented}
}
\value{
the object
}
\description{
Arg constructor yet to be implemented
}
\examples{
#no example yet
}
