% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Dockerfile.R
\docType{methods}
\name{as.character,Dockerfile-method}
\alias{as.character,Dockerfile-method}
\title{Convert a Dockerfile to String}
\usage{
\S4method{as.character}{Dockerfile}(x, ...)
}
\arguments{
\item{x}{Dockerfile.}

\item{...}{Arguments to be passed down to toString}
}
\description{
Convert a Dockerfile to String
}
