% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-class.R
\docType{methods}
\name{as.character,Instruction-method}
\alias{as.character,Instruction-method}
\title{Convert an Instruction-object to a string holding a Docker instruction}
\usage{
\S4method{as.character}{Instruction}(x)
}
\arguments{
\item{x}{Instruction}
}
\value{
A single character string
}
\description{
Convert an Instruction-object to a string holding a Docker instruction
}

