% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-All.R
\docType{class}
\name{characterOrInteger-class}
\alias{characterOrInteger-class}
\title{Class union for slots taking either numeric or character input}
\value{
object
}
\description{
Class union for slots taking either numeric or character input
}
