% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{clean_session}
\alias{clean_session}
\title{Obtains a \code{sessionInfo} from a local R session}
\usage{
clean_session(expr = c(), file = NULL, rmd_file = NULL,
  echo = FALSE, predetect = TRUE,
  repos = "http://cloud.r-project.org", cmd = Sys.getenv("R_HOME"))
}
\arguments{
\item{expr}{vector of expressions to be executed in the session (see \code{\link{quote}})}

\item{file}{R script to be executed in the session, uses \code{\link{source}}}

\item{echo}{print out detailed information from R}

\item{file_rmd}{R Markdown file to rendered in the session, uses \code{\link[rmarkdown]{render}}}

\item{pretedect}{whether to use \pkg{automagic} to install missing packaging before executing the R script or R Markdown file}
}
\description{
The function may also execute provided expressions or files.
}
\examples{
clean_session(c(quote(library('lattice'))))
}
