% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Cmd.R
\name{Cmd}
\alias{Cmd}
\title{create objects representing a CMD instruction}
\usage{
Cmd(exec = NA_character_, params = NA_character_, form = "exec")
}
\arguments{
\item{exec}{character argument naming the executable}

\item{params}{paramterer arguments}

\item{form}{form to render the instruction to (exec or shell)}
}
\value{
An S4 object of class \linkS4class{Cmd}
}
\description{
create objects representing a CMD instruction
}
\examples{
toString(Cmd("R", "--vanilla"))
}
