% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Cmd.R
\name{CMD_Render}
\alias{CMD_Render}
\title{Create CMD instruction for rendering a markdown file}
\usage{
CMD_Render(path, options = character(0),
  output_format = rmarkdown::html_document(), output_dir = NULL)
}
\arguments{
\item{path}{The name of the R markdown file that should run on startup or a path relative to the working directory}

\item{options}{(optional) Options or flags to be passed to Rscript}

\item{output_format}{The output format as in \code{rmakdown::render(...)}}

\item{outpur_dir}{The output dir as in \code{rmakdown::render(...)}}
}
\value{
A CMD instruction
}
\description{
Schema: Rscript [--options] -e \"rmarkdown::render(input = [path], output_format = [output_format]\""
}
\seealso{
\link{markdown::render}
}
