% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Cmd.R
\name{CMD_Rexpr}
\alias{CMD_Rexpr}
\title{Create CMD instruction for running an R expression}
\usage{
CMD_Rexpr(expr, options = character(0), args = character(0), vanilla = TRUE)
}
\arguments{
\item{expr}{character vector of expressions to run}

\item{options}{(optional) Options or flags to be passed to Rscript}

\item{args}{(otional) Argumands to be passed to the R script}

\item{vanilla}{Whether R should startup in vanilla mode. Default: TRUE}
}
\value{
A CMD instruction
}
\description{
Schema: R [--options] [file] [args]
}
