% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Cmd.R
\name{CMD_Rscript}
\alias{CMD_Rscript}
\title{Create CMD instruction for running an R script}
\usage{
CMD_Rscript(path, options = character(0), args = character(0), vanilla = TRUE)
}
\arguments{
\item{path}{The name of the R script that should run on startup or a path relative to the working directory}

\item{options}{(optional) Options or flags to be passed to Rscript}

\item{args}{(otional) Argumands to be passed to the R script}

\item{vanilla}{Whether R should startup in vanilla mode. Default: TRUE}
}
\value{
A CMD instruction
}
\description{
Schema: R [--options] [file] [args]
}
