% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{containerit-addin}
\alias{containerit-addin}
\title{Helper function for RStudio Addin user interface}
\description{
See \code{README.Rmd} for details about the RStudio Addin.
roxygen comment mainly to configure required imports for UI in one place.
}
