% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containerit.R
\docType{package}
\name{containerit}
\alias{containerit}
\alias{containers,}
\alias{containerization}
\alias{containerit-package}
\title{containerit: A package for packaging R objects and their dependencies in software containers}
\description{
The containerit package provides one core function: \code{dockerfile}.
}
\section{Create Dockerfiles}{

The \code{dockerfile} method create an R representation of a Dockerfile, which is a recipee
for a Docker container. The object can be serialized to a Dockerfile, which can be used to build
a runnable Docker image. \code{dockerfile} objects can be created based on R scripts and sessions.
}

