% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containerit-config.R
\name{containerit_write_config}
\alias{containerit_write_config}
\title{Re - init (overwrite) current package configuration}
\usage{
containerit_write_config(
  config = .getDefaultConfig(),
  output = file.path(find.package("containerit"), "inst/containerit_config.json")
)
}
\arguments{
\item{config}{The configuration object to be written}

\item{output}{Path or connection to write the configuration to}
}
\description{
Re - init (overwrite) current package configuration
}
\examples{
containerit:::containerit_write_config(output = stdout())
\dontrun{
#developers may want to initialize the default config with:
containerit:::containerit_write_config(output = "inst/containerit_config.json")
}
}
