% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Copy.R
\docType{class}
\name{Copy-class}
\alias{Copy-class}
\title{S4 Class representing a COPY-instruction}
\arguments{
\item{src}{(character vector) list of files or directories to be copied}

\item{dest}{(character string) destination directory on the docker image (either absolute or relative to working directory)}
}
\value{
object
}
\description{
S4 Class representing a COPY-instruction
}
\examples{
#no example yet
}

