% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Copy.R
\name{Copy}
\alias{Copy}
\title{Copy one or more files or directories to a Docker image}
\usage{
Copy(src, dest)
}
\arguments{
\item{src}{(character vector) list of files or directories to be copied}

\item{dest}{(character string) destination directory on the docker image (either absolute or relative to working directory)}
}
\value{
the object
}
\description{
Copy one or more files or directories to a Docker image
}
\examples{
#no example yet
}
