% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessionInfo-localbuild-methods.R
\name{docker_inspect}
\alias{docker_inspect}
\title{Read labels from images and containers}
\usage{
docker_inspect(host = harbor::localhost, name, docker_opts = character(0),
  ...)
}
\arguments{
\item{host}{A host object, as specified by the harbor package}

\item{name}{Name or id of the docker object (can also be a list of multiple names/ids)}

\item{docker_opts}{Options to docker. These are things that come before the docker command, when run on the command line. (as in harbor::docker:cmd)}

\item{...}{Other arguments passed to the SSH command for the host}
}
\value{
A named list of labels for each name or id given. (So, if there are multiple names/ids a list of named lists is returned)
}
\description{
See https://docs.docker.com/engine/reference/commandline/inspect/
}
\details{
The imlementation is based on \link[harbor]{docker_cmd} in the harbor package
}
\examples{
\dontrun{
 docker_inspect(name="rocker/r-ver:3.3.2")
}

}
\seealso{
\link[harbor]{docker_cmd}
}
