% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-class.R
\docType{class}
\name{Dockerfile-class}
\alias{Dockerfile}
\alias{Dockerfile-class}
\title{An S4 class to represent a Dockerfile}
\value{
an object of class \code{Dockerfile}
}
\description{
See official documentation at \url{https://docs.docker.com/engine/reference/builder/}.
}
\section{Slots}{

\describe{
\item{\code{image}}{the base image, used in the FROM statement (character)}

\item{\code{maintainer}}{the MAINTAINER (object of class maintainer)}

\item{\code{instructions}}{an ordered list of instructions in the Dockerfile (list of character)}

\item{\code{context}}{Directories that shall be included in the context when dockerfile is build}

\item{\code{cmd}}{the default cmd instruction applied to the container}
}}

