% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{empty_session}
\alias{empty_session}
\title{Creates an empty R session}
\usage{
empty_session()
}
\value{
An object of class \code{sessionInfo}
}
\description{
Creates an empty R session
}
\details{
Uses \code{\link{clean_session}}
}
