% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Entrypoint.R
\docType{class}
\name{Entrypoint-class}
\alias{Entrypoint-class}
\title{Instruction class yet to be implemented}
\value{
object
}
\description{
Instruction class yet to be implemented
}
\examples{
#no example yet
}
