% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Env.R
\name{Env}
\alias{Env}
\title{create objects of class Env}
\usage{
Env(argument, value = NULL)
}
\arguments{
\item{argument}{the argument name}

\item{value}{the value to be set to the argument}
}
\value{
Env-object
}
\description{
create objects of class Env
}
