% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{extract_session_file}
\alias{extract_session_file}
\title{Reads a \code{sessionInfo} object from an \code{Rdata} file}
\usage{
extract_session_file(file)
}
\arguments{
\item{file}{file path}
}
\value{
An object of class \code{sessionInfo}
}
\description{
Reads a \code{sessionInfo} object from an \code{Rdata} file
}
\examples{
sessionInfo <- sessionInfo()
file <- tempfile(tmpdir = tempdir(), fileext = ".Rdata")
save(sessionInfo, file = file)
extract_session_file(file)

}
