% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Dockerfile.R
\name{format,Dockerfile-method}
\alias{format,Dockerfile-method}
\title{Format a Dockerfile}
\usage{
\S4method{format}{Dockerfile}(x, ...)
}
\arguments{
\item{x}{Dockerfile}

\item{...}{Arguments to be passed down to format.default}
}
\description{
Format a Dockerfile
}
