% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile.R
\name{format.Dockerfile}
\alias{format.Dockerfile}
\title{Format a dockerfile object to a series of instruction}
\usage{
\method{format}{Dockerfile}(x, ...)
}
\arguments{
\item{x}{An object of class Dockerfile}

\item{...}{Arguments to be passed down to format.default}
}
\value{
The Content of the dockerfile represented by the dockerfile object, by default formated as a List of strings where each string represent a new line
}
\description{
Format a dockerfile object to a series of instruction
}
\examples{
format(dockerfile())
}

