% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-class.R
\name{From}
\alias{From}
\title{create objects of class From}
\usage{
From(image, tag = NULL, digest = NULL)
}
\arguments{
\item{image}{image name or image-id}

\item{tag}{optional image tag (character)}

\item{digest}{optional image digest (character)}
}
\value{
From-object
}
\description{
create objects of class From
}

