% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{get_docker_cran_mirror}
\alias{get_docker_cran_mirror}
\title{Get default cran mirror to be used to install R packages}
\usage{
get_docker_cran_mirror()
}
\value{
The mirror's url as single character string (default: "https://cloud.r-project.org")
}
\description{
Get default cran mirror to be used to install R packages
}

