% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile.R
\name{getImageForVersion}
\alias{getImageForVersion}
\title{getImageForVersion-method}
\usage{
getImageForVersion(r_version)
}
\arguments{
\item{r_version}{A string representation of the R version, e.g. "3.4.2"}
}
\value{
A string with the name of the Docker image
}
\description{
Get a suitable Rocker image based on the R version.
}
\examples{
getImageForVersion(getRVersionTag(utils::sessionInfo()))
getImageForVersion("3.4.2")

}
