% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Healthcheck.R
\name{Healthcheck}
\alias{Healthcheck}
\title{Healthcheck constructor yet to be implemented}
\usage{
Healthcheck(...)
}
\arguments{
\item{...}{fields yet to be implemented}
}
\value{
the object
}
\description{
Healthcheck constructor yet to be implemented
}
\examples{
#no example yet
}
