% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Instruction.R
\docType{class}
\name{Instruction-class}
\alias{Instruction-class}
\alias{Instruction}
\title{The Docker Instruction - Class}
\description{
See official documentation at \url{https://docs.docker.com/engine/reference/builder/#format}.
}
