% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Label.R
\docType{class}
\name{Label-class}
\alias{Label-class}
\title{Instruction class representing a LABEL}
\value{
object
}
\description{
See Dockerfile reference at https://docs.docker.com/engine/reference/builder/#label
See project issue at https://github.com/o2r-project/containerit/issues/42
}
\seealso{
Other label: \code{\link{LabelSchemaFactory}},
  \code{\link{Label_Maintainer}},
  \code{\link{Label_SessionInfo}}, \code{\link{Label}}

Other instruction classes: \code{\link{Add-class}},
  \code{\link{Cmd-class}}, \code{\link{Copy-class}},
  \code{\link{Entrypoint-class}}, \code{\link{Env-class}},
  \code{\link{Expose-class}}, \code{\link{From-class}},
  \code{\link{Instruction-class}},
  \code{\link{Maintainer-class}}, \code{\link{Run-class}},
  \code{\link{Run_shell-class}},
  \code{\link{Volume-class}}, \code{\link{Workdir-class}}
}
