% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Label.R
\name{Label}
\alias{Label}
\title{Creates a label of arbitrary content}
\usage{
Label(..., multi_line = FALSE)
}
\arguments{
\item{...}{One or more named arguments (key-value pairs), that the Instruction should consist of}

\item{multi_line}{Whether to print labels with multiple values in one single line or to put each key-value pair in a new line.}
}
\value{
the object
}
\description{
See https://docs.docker.com/engine/reference/builder/#label
}
\examples{
#no example yet
}
