% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Label.R
\name{Label_Maintainer}
\alias{Label_Maintainer}
\title{Create Label corresponding to the maintainer field}
\usage{
Label_Maintainer(value)
}
\arguments{
\item{value}{Name or email of the maintainer, see specification}
}
\value{
An object of class 'maintainer'
}
\description{
See https://docs.docker.com/engine/reference/builder/#maintainer-deprecated
}
\examples{
Label_Maintainer("SvenDowideit@home.org.au")

}
