% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Label.R
\name{Label_SessionInfo}
\alias{Label_SessionInfo}
\title{Creates a label holding R session information}
\usage{
Label_SessionInfo(session = sessionInfo(), as_json = FALSE)
}
\arguments{
\item{session}{An object of class 'sessionInfo' as returned by utils::sessionInfo(),
or alternatively a class 'session_info' as returned by devtools::session_info()}

\item{as_json}{Determines whether to parse the session information to a json-string instead of a plain string}
}
\value{
A label with key 'R.session-info' and the deparsed session information in one line
}
\description{
See https://docs.docker.com/engine/reference/builder/#label
}
\examples{
the_session <- clean_session()
the_dockerfile <- dockerfile(the_session)
addInstruction(the_dockerfile) <- Label_SessionInfo(the_session)
}
\seealso{
Other label: 
\code{\link{Label-class}},
\code{\link{LabelSchemaFactory}()},
\code{\link{Label_Maintainer}()},
\code{\link{Label}}
}
\concept{label}
