% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-class.R
\docType{class}
\name{Maintainer-class}
\alias{Maintainer}
\alias{Maintainer-class}
\title{An S4 class to represent a Dockerfile's maintainer}
\value{
an object of class \code{Maintainer}
}
\description{
An S4 class to represent a Dockerfile's maintainer
}
\section{Slots}{

\describe{
\item{\code{name}}{The name}

\item{\code{email}}{The email}
}}

