% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Maintainer.R
\name{Maintainer}
\alias{Maintainer}
\title{Create MAINTAINER instructions}
\usage{
Maintainer(name, email = NA_character_)
}
\arguments{
\item{name}{character name (by convention: "<given name> <last name>", e.g. "Matthias Hinz")}

\item{email}{The email}
}
\value{
Maintainer object
}
\description{
The MAINTAINER instruction is deprecated!
}
\seealso{
Other maintainer: 
\code{\link{Label_Maintainer}()},
\code{\link{Maintainer-class}}
}
\concept{maintainer}
