% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Onbuild.R
\name{Onbuild}
\alias{Onbuild}
\title{Onbuild constructor yet to be implemented}
\usage{
Onbuild(...)
}
\arguments{
\item{...}{fields yet to be implemented}
}
\value{
the object
}
\description{
Onbuild constructor yet to be implemented
}
\examples{
#no example yet
}
