% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Instruction.R
\name{print,Instruction-method}
\alias{print,Instruction-method}
\title{Print an Instruction}
\usage{
\S4method{print}{Instruction}(x, ...)
}
\arguments{
\item{x}{Instruction.}

\item{...}{Arguments to be passed down to toString}
}
\description{
Print an Instruction
}
