% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-class.R
\name{Run}
\alias{Run}
\title{create objects representing a RUN instruction}
\usage{
Run(exec, params = NA_character_)
}
\arguments{
\item{exec}{character argument naming the executable}

\item{params}{paramterer arguments}
}
\value{
An S4 object of class Run
}
\description{
create objects representing a RUN instruction
}

