% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Run.R
\docType{class}
\name{Run_shell-class}
\alias{Run_shell-class}
\title{An S4 class to represent a RUN instruction in shell form}
\value{
object of class Run
}
\description{
An S4 class to represent a RUN instruction in shell form
}
\section{Slots}{

\describe{
\item{\code{commands}}{}
}}

