% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Run.R
\name{Run_shell}
\alias{Run_shell}
\title{Create objects representing a RUN instruction in shell form}
\usage{
Run_shell(commands)
}
\arguments{
\item{commands}{character vector of commands (will be concatenated with && and linebreaks)}
}
\value{
An S4 object of class Run_shell
}
\description{
Create objects representing a RUN instruction in shell form
}
\seealso{
Other Run instruction: \code{\link{Run-class}},
  \code{\link{Run_shell-class}}, \code{\link{Run}}
}
\concept{Run instruction}
