% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{set_docker_cran_mirror}
\alias{set_docker_cran_mirror}
\title{Set default cran mirror to be used to install R packages}
\usage{
set_docker_cran_mirror(cran_url)
}
\arguments{
\item{cran_url}{The mirror's url as single character string}
}
\description{
Set default cran mirror to be used to install R packages
}

