% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Shell.R
\name{Shell}
\alias{Shell}
\title{Shell constructor yet to be implemented}
\usage{
Shell(...)
}
\arguments{
\item{...}{fields yet to be implemented}
}
\value{
the object
}
\description{
Shell constructor yet to be implemented
}
\examples{
#no example yet
}
