% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Stopsignal.R
\name{Stopsignal}
\alias{Stopsignal}
\title{Stopsignal constructor yet to be implemented}
\usage{
Stopsignal(...)
}
\arguments{
\item{...}{fields yet to be implemented}
}
\value{
the object
}
\description{
Stopsignal constructor yet to be implemented
}
\examples{
#no example yet
}
