% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Dockerfile.R
\docType{methods}
\name{toString,Dockerfile-method}
\alias{toString,Dockerfile-method}
\title{Convert a Dockerfile to String}
\usage{
\S4method{toString}{Dockerfile}(x, ...)
}
\arguments{
\item{x}{Dockerfile.}

\item{...}{Other params (not in use currently)}
}
\description{
Convert a Dockerfile to String
}
