% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Volume.R
\name{Volume}
\alias{Volume}
\title{Constructor yet to be implemented}
\usage{
Volume(...)
}
\arguments{
\item{...}{fields yet to be implemented}
}
\value{
the object
}
\description{
Constructor yet to be implemented
}
\examples{
#no example yet
}
