% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Workdir.R
\docType{class}
\name{Workdir-class}
\alias{Workdir-class}
\title{Instruction class yet to be implemented}
\value{
object
}
\description{
Instruction class yet to be implemented
}
\examples{
#no example yet
}

