% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Workdir.R
\name{Workdir}
\alias{Workdir}
\title{Constructor yet to be implemented}
\usage{
Workdir(...)
}
\arguments{
\item{...}{fields yet to be implemented}
}
\value{
the object
}
\description{
Constructor yet to be implemented
}
\examples{
#no example yet
}
