% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile.R
\docType{methods}
\name{write,Dockerfile-method}
\alias{write,Dockerfile-method}
\title{Write a Dockerfile object to a dockerfile}
\usage{
\S4method{write}{Dockerfile}(x, file = paste0(getwd(), "/", "Dockerfile"))
}
\arguments{
\item{x}{Dockerfile object to be serialized}

\item{file}{optional argument specifying a costum file path}
}
\description{
Write a Dockerfile object to a dockerfile
}
\examples{
# write a dockerfile with default parameters to temporary file and show content:
temp = tempfile()
write(dockerfile(), file=temp) 
print(readLines(temp)) 
unlink(temp)
}

