% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile.R
\name{write.Dockerfile}
\alias{write.Dockerfile}
\title{Write a dockerfile}
\usage{
write.Dockerfile(x, file = paste0(getwd(), "/", "Dockerfile"))
}
\arguments{
\item{x}{Dockerfile-object to be serialized}

\item{file}{optional argument for specifying a costum file path}
}
\description{
Write a dockerfile
}
\examples{
# write a dockerfile with default parameters to temporary file and show content:
temp = tempfile()
write.Dockerfile(dockerfile(), file=temp) 
print(readLines(temp)) 
unlink(temp)
}

