% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Arg.R
\name{Arg}
\alias{Arg}
\title{Create objects representing a ARG instruction}
\usage{
Arg(key, value = NA_character_)
}
\arguments{
\item{key}{character argument naming the argument}

\item{value}{default value for the argument, of class character}
}
\value{
An S4 object of class Arg
}
\description{
Create objects representing a ARG instruction
}
\seealso{
Other Arg instruction: 
\code{\link{Arg-class}}
}
\concept{Arg instruction}
