% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Instruction.R
\name{as.character,Instruction-method}
\alias{as.character,Instruction-method}
\title{Convert an Instruction-object to a string holding a Docker instruction}
\usage{
\S4method{as.character}{Instruction}(x, ...)
}
\arguments{
\item{x}{Instruction object (of class Run, Cmd, From ...)}

\item{...}{Arguments to be passed down to toString}
}
\value{
A single character string in Dockerfile syntax
}
\description{
Convert an Instruction-object to a string holding a Docker instruction
}
