% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Cmd.R
\docType{class}
\name{Cmd-class}
\alias{Cmd-class}
\title{An S4 class to represent a CMD instruction}
\value{
object of class cmd
}
\description{
An S4 class to represent a CMD instruction
}
\section{Slots}{

\describe{
\item{\code{exec}}{character.}

\item{\code{params}}{character.}
}}

