% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Comment.R
\name{Comment}
\alias{Comment}
\title{Creates a comment with character content}
\usage{
Comment(text)
}
\arguments{
\item{text}{The comment text}
}
\value{
the object
}
\description{
Creates a comment with character content
}
\examples{
# A simple label that occupies one line:
comment <- Comment(text = "this is important to note")
the_dockerfile <- dockerfile(clean_session())
addInstruction(the_dockerfile) <- comment
print(the_dockerfile)

}
