% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containerit.R
\docType{package}
\name{containerit}
\alias{containerit}
\alias{containers,}
\alias{containerization}
\alias{containerit-package}
\title{containerit: A package for packaging R objects and their dependencies in software containers}
\description{
The containerit package provides two core functions:
Dockerfile, and container.
}
\section{Dockerfile functions}{

The dockerfile functions create a recipee for a Docker container based on R scripts or objects.
}

\section{container functions}{

The container functions create the actual container wrapping calls to a local Docker CLI.
}

