% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containeRit.R
\docType{package}
\name{containeRit}
\alias{containeRit}
\alias{containeRit-package}
\alias{containerization}
\alias{containers,}
\title{containeRit: A package for packaging R objects and their dependencies in software containers}
\description{
The containeRit package provides two core functions:
Dockerfile, and container.
}
\section{Dockerfile functions}{

The dockerfile functions create a recipee for a Docker container based on R scripts or objects.
}

\section{container functions}{

The container functions create the actual container wrapping calls to a local Docker CLI.
}

