% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containeRit-config.R
\name{containeRit_write_config}
\alias{containeRit_write_config}
\title{Re - init (overwrite) current package configuration}
\usage{
containeRit_write_config(config = .getDefaultConfig(),
  output = file.path(find.package("containeRit"), "containeRit_config.json"))
}
\arguments{
\item{config}{The configuration object to be written}

\item{output}{Path or connection to write the configuration to}
}
\description{
Re - init (overwrite) current package configuration
}
\examples{
containeRit_write_config(output = stdout())
\dontrun{
#developers may want to initialize the default config with:
containeRit_write_config(output = "inst/containeRit_config.json")
}
}
