% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Copy.R
\name{Copy}
\alias{Copy}
\title{Constructor yet to be implemented}
\usage{
Copy(...)
}
\arguments{
\item{...}{fields yet to be implemented}
}
\value{
the object
}
\description{
Constructor yet to be implemented
}
\examples{
#no example yet
}
