% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{docker_build}
\alias{docker_build}
\title{Build a Docker image from a local Dockerfile}
\usage{
docker_build(context, tag, the_dockerfile = "Dockerfile", ...)
}
\arguments{
\item{context}{Local location of build directory including valid Dockerfile}

\item{tag}{Name of the new image to be created}

\item{...}{Other arguments passed to stevedore}

\item{dockerfile}{(optional) set path to the dockerfile (equals to path/to/Dockerfile"))}
}
\value{
A table of active images on the instance
}
\description{
Uploads a folder with a Dockerfile and supporting files to an instance and builds it.
The method is implemented based on \pkg{stevedore}.
}
