% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{docker_build}
\alias{docker_build}
\title{Build a Docker image from a local Dockerfile}
\usage{
docker_build(host = harbor::localhost, dockerfolder, tag,
  dockerfile = character(0), wait = FALSE, no_cache = FALSE,
  docker_opts = character(0), ...)
}
\arguments{
\item{host}{A host object (see harbor-package)}

\item{dockerfolder}{Local location of build directory including valid Dockerfile}

\item{tag}{Name of the new image to be created}

\item{dockerfile}{(optional) set path to the dockerfile (equals to path/to/Dockerfile"))}

\item{wait}{Whether to block R console until finished build}

\item{no_cache}{Wheter to use cached layers to build the image}

\item{docker_opts}{Additional Docker options}

\item{...}{Other arguments passed to the SSH command for the host}
}
\value{
A table of active images on the instance
}
\description{
Uploads a folder with a Dockerfile and supporting files to an instance and builds it.
The method is implemented based on \code{harbor::docker_cmd} and analogue to \code{googleComputeEngineR::docker_build} (with small differences).
}
