% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Dockerfile.R
\docType{class}
\name{Dockerfile-class}
\alias{Dockerfile-class}
\alias{Dockerfile}
\title{An S4 class to represent a Dockerfile}
\value{
an object of class \code{Dockerfile}
}
\description{
An S4 class to represent a Dockerfile
}
\section{Slots}{

\describe{
\item{\code{image}}{the base image, used in the FROM statement (character)}

\item{\code{maintainer}}{a maintainer label or an object of class maintainer (the latter is debrecated)}

\item{\code{instructions}}{an ordered list of instructions in the Dockerfile (list of character)}

\item{\code{cmd}}{the default cmd instruction applied to the container}
}}

