% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-class.R
\docType{class}
\name{Dockerfile-class}
\alias{Dockerfile}
\alias{Dockerfile-class}
\title{An S4 class to represent a Dockerfile}
\value{
an object of class \code{Dockerfile}
}
\description{
See official documentation at \url{https://docs.docker.com/engine/reference/builder/}.

See official documentation at \url{https://docs.docker.com/engine/reference/builder/}.
}
\section{Slots}{

\describe{
\item{\code{image}}{the base image, used in the FROM statement}

\item{\code{maintainer}}{the MAINTAINER}

\item{\code{instructions}}{an ordered list of instructions in the Dockerfile}

\item{\code{from}}{the base image}

\item{\code{instructions}}{an ordered list of instructions in the Dockerfile}
}}

