% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile.R
\name{dockerfile}
\alias{dockerfile}
\title{dockerfile}
\usage{
dockerfile(from = utils::sessionInfo(), env = NULL, maintainer = NULL,
  r_version = paste(R.Version()$major, R.Version()$minor, sep = "."),
  image = NULL, context = NA_character_)
}
\arguments{
\item{from}{the source of the information to construct the Dockerfile}

\item{env}{the environment that should be included in the image}

\item{maintainer}{optionally specify the maintainer of the dockerfile (see Maintainter-class)}

\item{r_version}{(character) optionally specify the R version that should run inside the container (default: the current R version that runs on host)
containeRit will try to find a corresponding docker image as a base (if not sepcified otherwise)}

\item{image}{(character) optionally specify the image that shall be used for the docker container (FROM-statement)}

\item{context}{(character vector) optionally specify one or many build context paths}
}
\value{
An object of class Dockerfile
}
\description{
Create a Dockerfile from the
}
\examples{
dockerfile()

}

