% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile.R
\name{dockerfile}
\alias{dockerfile}
\title{dockerfile}
\usage{
dockerfile(from = utils::sessionInfo(), to = paste0(getwd(), "/",
  "Dockerfile"), env = NULL, maintainer = NULL,
  r_version = paste(R.Version()$major, R.Version()$minor, sep = "."))
}
\arguments{
\item{from}{the source of the information to construct the Dockerfile}

\item{to}{path and file name to save the Dockerfile to}

\item{env}{the environment that should be included in the image}

\item{maintainer}{optionally specify the maintainer of the dockerfile (see Maintainter-class)}

\item{r_version}{optionally specify the R version that should run inside the container (uses the current version that runs on host, by default)}
}
\value{
An object of class Dockerfile
}
\description{
Create a Dockerfile from the
}
\examples{
dockerfile()

}

