% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile.R
\name{dockerfile}
\alias{dockerfile}
\title{dockerfile}
\usage{
dockerfile(from = utils::sessionInfo(), to = paste0(getwd(), "/",
  "Dockerfile"), env = NULL)
}
\arguments{
\item{from}{the source of the information to construct the Dockerfile}

\item{to}{path and file name to save the Dockerfile to}

\item{env}{the environment that should be included in the image}
}
\value{
An object of class Dockerfile
}
\description{
Create a Dockerfile from the
}
\examples{
dockerfile()

}

