% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Entrypoint.R
\docType{class}
\name{Entrypoint-class}
\alias{Entrypoint-class}
\title{Instruction class yet to be implemented}
\value{
object
}
\description{
Instruction class yet to be implemented
}
\examples{
#no example yet
}
\seealso{
Other instruction classes: \code{\link{Add-class}},
  \code{\link{Cmd-class}}, \code{\link{Copy-class}},
  \code{\link{Env-class}}, \code{\link{Expose-class}},
  \code{\link{From-class}},
  \code{\link{Instruction-class}},
  \code{\link{Label-class}},
  \code{\link{Maintainer-class}}, \code{\link{Run-class}},
  \code{\link{Run_shell-class}},
  \code{\link{Volume-class}}, \code{\link{Workdir-class}}
}
