% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Entrypoint.R
\name{Entrypoint}
\alias{Entrypoint}
\title{Constructor yet to be implemented}
\usage{
Entrypoint(...)
}
\arguments{
\item{...}{fields yet to be implemented}
}
\value{
the object
}
\description{
Constructor yet to be implemented
}
\examples{
#no example yet
}

