% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Entrypoint.R
\name{Entrypoint}
\alias{Entrypoint}
\title{Constructor for Entrypoint yet to be implemented}
\usage{
Entrypoint(program, params = list(), form = "exec")
}
\arguments{
\item{program}{the executable or command}

\item{params}{the parameters provided to the executable or command}

\item{form}{the form how the instruction is rendered, either \code{exec} or \code{shell}}
}
\value{
an object of class \linkS4class{Entrypoint}
}
\description{
Constructor for Entrypoint yet to be implemented
}
\examples{
toString(Entrypoint("Rscript", params = list("sum", "1", "2")))
toString(Entrypoint("/bin/echo"))
}
