% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Env.R
\docType{class}
\name{Env-class}
\alias{Env-class}
\title{Env-instruction class yet to be implemented}
\value{
object of class Env
}
\description{
Env-instruction class yet to be implemented
}
\section{Slots}{

\describe{
\item{\code{key}}{character.}

\item{\code{value}}{character.}
}}

\seealso{
Other instruction classes: 
\code{\link{Add-class}},
\code{\link{Arg-class}},
\code{\link{Cmd-class}},
\code{\link{Comment-class}},
\code{\link{Copy-class}},
\code{\link{Entrypoint-class}},
\code{\link{Expose-class}},
\code{\link{From-class}},
\code{\link{Healthcheck-class}},
\code{\link{Instruction-class}},
\code{\link{Label-class}},
\code{\link{Maintainer-class}},
\code{\link{Onbuild-class}},
\code{\link{Run-class}},
\code{\link{Run_shell-class}},
\code{\link{Shell-class}},
\code{\link{Stopsignal-class}},
\code{\link{Volume-class}},
\code{\link{Workdir-class}}

Other Env instruction: 
\code{\link{Env}}
}
\concept{Env instruction}
\concept{instruction classes}
