% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Env.R
\name{Env}
\alias{Env}
\title{Constructor for Env yet to be implemented}
\usage{
Env(...)
}
\arguments{
\item{...}{fields yet to be implemented}
}
\value{
the object
}
\description{
Constructor for Env yet to be implemented
}
\examples{
#no example yet
}
