% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Env.R
\name{Env}
\alias{Env}
\title{Create objects representing a ENV instruction}
\usage{
Env(key, value = NA_character_)
}
\arguments{
\item{key}{character argument naming the argument}

\item{value}{default value for the argument, of class character}
}
\value{
An S4 object of class Env
}
\description{
Create objects representing a ENV instruction
}
\seealso{
Other Env instruction: 
\code{\link{Env-class}}
}
\concept{Env instruction}
