% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Expose.R
\name{Expose}
\alias{Expose}
\title{Constructor yet to be implemented}
\usage{
Expose(...)
}
\arguments{
\item{...}{fields yet to be implemented}
}
\value{
the object
}
\description{
Constructor yet to be implemented
}
\examples{
#no example yet
}

