% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Expose.R
\name{Expose}
\alias{Expose}
\title{Constructor for Expose}
\usage{
Expose(port, host = NA_integer_)
}
\arguments{
\item{port}{(character or numeric, required) the port of the container to be exposed}

\item{host}{(numeric, optional) port of the host, if missing the same port as in the container is exposed}
}
\value{
the object
}
\description{
Constructor for Expose
}
\examples{
#no example yet
}
