% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile.R
\name{format.Dockerfile}
\alias{format.Dockerfile}
\title{Format a Dockerfile object to a series of instructions}
\usage{
\method{format}{Dockerfile}(x, ...)
}
\arguments{
\item{x}{An object of class Dockerfile}

\item{...}{Arguments to be passed down to format.default}
}
\value{
The content of the Dockerfile represented by the Dockerfile object, by default formatted as a list of strings where each string represent a new line
}
\description{
Format a Dockerfile object to a series of instructions
}
\examples{
format(dockerfile())
}

