% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-class.R
\docType{class}
\name{From-class}
\alias{From-class}
\title{An S4 class to represent a FROM instrunction relating to a docker image}
\value{
an object of class \code{Docker From}
}
\description{
An S4 class to represent a FROM instrunction relating to a docker image
}
\section{Slots}{

\describe{
\item{\code{image}}{image-id or image name}

\item{\code{postfix}}{tag or digest}
}}

