% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseimages.R
\name{get_installed_packages}
\alias{get_installed_packages}
\title{Detect packages in a Docker image}
\usage{
get_installed_packages(image)
}
\arguments{
\item{image}{Name of the new image to be run}
}
\value{
A \code{data.frame} with installed packages and their version of the image
}
\description{
Extracts all installed packages from a Docker image by running it.
The method is based on \code{stevedore}.
}
