% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-installation-methods.R
\name{getGitHubRef}
\alias{getGitHubRef}
\title{Get GitHub reference from package}
\usage{
getGitHubRef(pkg, pkgs = c(sessionInfo()$otherPkgs, sessionInfo()$loadedOnly))
}
\arguments{
\item{pkg}{The name of the package to retrieve the reference for}

\item{pkgs}{Lists of packages from a sessionInfo object}
}
\value{
A character string with a short refernce, e.g. \code{r-hub/sysreqs@481d263}, \code{NA} is nothign could be found
}
\description{
If a package is installed from GitHub this function tries to retrieve the reference (i.e. user name, repository name, and commit) from
(a) the provided sessionInfo, or
(b) locally, and only if the package is installed (!), using \code{\link[devtools]{session_info}}.
}
\examples{
\dontrun{
getGitHubRef(rsysreqs)
}

}
