% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile.R
\name{getRVersionTag}
\alias{getRVersionTag}
\title{Get R version from a variety of sources in a string format used for image tags}
\usage{
getRVersionTag(from = NULL, default = R.Version())
}
\arguments{
\item{from}{the source to extract an R version: an `sessionInfo()` object}
}
\description{
Returns either a version extracted from a given object or the default version.
}
\examples{
getRVersionTag(from = sessionInfo())
getRVersionTag()
}
