% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{getRVersionTag}
\alias{getRVersionTag}
\title{Get R version in string format used for image tags}
\usage{
getRVersionTag(from, default = paste(R.Version()$major,
  R.Version()$minor, sep = "."))
}
\arguments{
\item{from}{the source to extract an R version: a `sessionInfo()` or `session_info()` object, or an \code{RData} file with a session info object}

\item{default}{if 'from' does not contain version information (e.g. its an Rscript), use this default version information.}
}
\description{
Returns either a version extracted from a given object or the default version.
}
\examples{
getRVersionTag(from = sessionInfo())
}
