% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Instruction.R
\docType{class}
\name{Instruction-class}
\alias{Instruction-class}
\alias{Instruction}
\title{The Docker Instruction - Class}
\description{
See official documentation at \url{https://docs.docker.com/engine/reference/builder/#format}.
}
\seealso{
Other instruction classes: \code{\link{Add-class}},
  \code{\link{Arg-class}}, \code{\link{Cmd-class}},
  \code{\link{Comment-class}}, \code{\link{Copy-class}},
  \code{\link{Entrypoint-class}}, \code{\link{Env-class}},
  \code{\link{Expose-class}}, \code{\link{From-class}},
  \code{\link{Healthcheck-class}},
  \code{\link{Label-class}},
  \code{\link{Maintainer-class}},
  \code{\link{Onbuild-class}}, \code{\link{Run-class}},
  \code{\link{Run_shell-class}}, \code{\link{Shell-class}},
  \code{\link{Stopsignal-class}},
  \code{\link{Volume-class}}, \code{\link{Workdir-class}}
}
\concept{instruction classes}
