% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Label.R
\docType{class}
\name{Label-class}
\alias{Label-class}
\title{Instruction class representing a LABEL}
\value{
object
}
\description{
See Dockerfile reference at https://docs.docker.com/engine/reference/builder/#label
See project issue at https://github.com/o2r-project/containerit/issues/42
}
