% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Label.R
\name{Label}
\alias{Label}
\title{Constructor yet to be implemented}
\usage{
Label(...)
}
\arguments{
\item{...}{fields yet to be implemented}
}
\value{
the object
}
\description{
Constructor yet to be implemented
}
\examples{
#no example yet
}

