% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Maintainer.R
\docType{class}
\name{Maintainer-class}
\alias{Maintainer-class}
\title{An S4 class to represent a Dockerfile's maintainer}
\value{
an object of class \code{Maintainer}
}
\description{
An S4 class to represent a Dockerfile's maintainer
}
\section{Slots}{

\describe{
\item{\code{name}}{The name}

\item{\code{email}}{The email}
}}

\seealso{
Other instruction classes: 
\code{\link{Add-class}},
\code{\link{Arg-class}},
\code{\link{Cmd-class}},
\code{\link{Comment-class}},
\code{\link{Copy-class}},
\code{\link{Entrypoint-class}},
\code{\link{Env-class}},
\code{\link{Expose-class}},
\code{\link{From-class}},
\code{\link{Healthcheck-class}},
\code{\link{Instruction-class}},
\code{\link{Label-class}},
\code{\link{Onbuild-class}},
\code{\link{Run-class}},
\code{\link{Run_shell-class}},
\code{\link{Shell-class}},
\code{\link{Stopsignal-class}},
\code{\link{Volume-class}},
\code{\link{Workdir-class}}

Other maintainer: 
\code{\link{Label_Maintainer}()},
\code{\link{Maintainer}}
}
\concept{instruction classes}
\concept{maintainer}
