% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Maintainer.R
\name{Maintainer}
\alias{Maintainer}
\title{create objects of class Maintainer}
\usage{
Maintainer(name, email = NA_character_)
}
\arguments{
\item{name}{character name (by convention: "<given name> <last name>", e.g. "Matthias Hinz")}

\item{email}{The email}
}
\value{
Maintainer object
}
\description{
create objects of class Maintainer
}
\seealso{
Other maintainer: \code{\link{Label_Maintainer}},
  \code{\link{Maintainer-class}}
}
