% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-class.R
\name{parseFrom}
\alias{parseFrom}
\title{Parse a From-instruction from an image-argument}
\usage{
parseFrom(string)
}
\arguments{
\item{string}{specifying ans image, i.e. <image-id>, <image@digest>, image:tag}
}
\value{
From-object
}
\description{
Parse a From-instruction from an image-argument
}

