% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-From.R
\name{parseFrom}
\alias{parseFrom}
\title{Parse a From-instruction from an image-argument}
\usage{
parseFrom(string)
}
\arguments{
\item{string}{Single character string that specifies an image in docker-syntax, e.g. \code{image}, \code{image@digest}, \code{image:tag}}
}
\value{
From instruction object
}
\description{
Parse a From-instruction from an image-argument
}
