% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-class.R
\docType{class}
\name{Run-class}
\alias{Run-class}
\title{An S4 class to represent a RUN instruction}
\value{
object of class Run
}
\description{
An S4 class to represent a RUN instruction
}
\section{Slots}{

\describe{
\item{\code{exec}}{character.}

\item{\code{params}}{character.}
}}

