% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Run.R
\name{Run}
\alias{Run}
\title{Create objects representing a RUN instruction}
\usage{
Run(exec, params = NA_character_)
}
\arguments{
\item{exec}{character argument naming the executable}

\item{params}{parameter arguments}
}
\value{
An S4 object of class Run
}
\description{
Create objects representing a RUN instruction
}
\seealso{
Other Run instruction: \code{\link{Run-class}},
  \code{\link{Run_shell-class}}, \code{\link{Run_shell}}
}
\concept{Run instruction}
