% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Run.R
\docType{class}
\name{Run_shell-class}
\alias{Run_shell-class}
\title{An S4 class to represent a RUN instruction in shell form}
\value{
object of class Run
}
\description{
An S4 class to represent a RUN instruction in shell form
}
\section{Slots}{

\describe{
\item{\code{commands}}{One or more commands to be included in the label}
}}

\seealso{
Other instruction classes: \code{\link{Add-class}},
  \code{\link{Arg-class}}, \code{\link{Cmd-class}},
  \code{\link{Copy-class}}, \code{\link{Entrypoint-class}},
  \code{\link{Env-class}}, \code{\link{Expose-class}},
  \code{\link{From-class}},
  \code{\link{Healthcheck-class}},
  \code{\link{Instruction-class}},
  \code{\link{Label-class}},
  \code{\link{Maintainer-class}},
  \code{\link{Onbuild-class}}, \code{\link{Run-class}},
  \code{\link{Shell-class}},
  \code{\link{Stopsignal-class}},
  \code{\link{Volume-class}}, \code{\link{Workdir-class}}

Other Run instruction: \code{\link{Run-class}},
  \code{\link{Run_shell}}, \code{\link{Run}}
}
