% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-installation-methods.R
\name{versioned_install_instructions}
\alias{versioned_install_instructions}
\title{helper function, expects a data.frame with columns name and version}
\usage{
versioned_install_instructions(pkgs)
}
\description{
helper function, expects a data.frame with columns name and version
}
