% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile-class.R
\docType{class}
\name{Volume-class}
\alias{Volume-class}
\title{Instruction class yet to be implemented}
\value{
object
}
\description{
Instruction class yet to be implemented
}
\examples{
#no example yet
}

