% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Workdir.R
\docType{class}
\name{Workdir-class}
\alias{Workdir-class}
\title{S4 Class representing a WORKDIR instruction}
\value{
object
}
\description{
S4 Class representing a WORKDIR instruction
}
\examples{
#no example yet
}
