% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Workdir.R
\name{Workdir}
\alias{Workdir}
\title{Constructor for a WORKDIR instruction}
\usage{
Workdir(path)
}
\arguments{
\item{path}{The path of the working directory}
}
\value{
the object
}
\description{
Constructor for a WORKDIR instruction
}
\examples{
instruction <- containerit::Workdir("~/myDir/subdir/")
toString(instruction)
}
